Ensure user docs use consistent ID
