+++ app/api/avatar/route.ts
@@ -1,14 +1,7 @@
 import { NextResponse, type NextRequest } from "next/server";
 import { currentUser } from "@clerk/nextjs/server";
-import { createClient } from "next-sanity";
-
-const client = createClient({
-  projectId: "8n5iznjt",
-  dataset: "production",
-  apiVersion: "2025-06-09",
-  token: process.env.SANITY_API_TOKEN,
-  useCdn: false,
-});
+import { client } from "@/app/sanity/client";
+import { ensureUser } from "@/app/sanity/user";
 
 export async function POST(req: NextRequest) {
   const user = await currentUser();
@@ -24,14 +17,20 @@ export async function POST(req: NextRequest) {
     contentType: file.type,
   });
 
+  const docId = await ensureUser({
+    id: user.id,
+    email: user.primaryEmailAddress?.emailAddress,
+    fullName: user.fullName,
+  });
+
   const profileId = await client.fetch(
     '*[_type=="profile" && user._ref==$id][0]._id',
-    { id: user.id }
+    { id: docId }
   );
   if (!profileId) {
     const created = await client.create({
       _type: "profile",
-      user: { _type: "reference", _ref: user.id },
+      user: { _type: "reference", _ref: docId },
       avatar: {
         _type: "image",
         asset: { _type: "reference", _ref: asset._id },
@@ -39,7 +38,6 @@ export async function POST(req: NextRequest) {
     });
     return NextResponse.json({ profile: created });
   }
-  if (!profileId) return new NextResponse("Profile not found", { status: 404 });
 
   const updated = await client
     .patch(profileId)
