+++ app/api/profile/route.ts
@@ -1,14 +1,7 @@
 import { NextResponse, type NextRequest } from "next/server";
 import { currentUser } from "@clerk/nextjs/server";
-import { createClient } from "next-sanity";
-
-const client = createClient({
-  projectId: "8n5iznjt",
-  dataset: "production",
-  apiVersion: "2025-06-09",
-  token: process.env.SANITY_API_TOKEN,
-  useCdn: false,
-});
+import { client } from "@/app/sanity/client";
+import { ensureUser } from "@/app/sanity/user";
 
 export async function POST(req: NextRequest) {
   const user = await currentUser();
@@ -18,18 +11,15 @@ export async function POST(req: NextRequest) {
 
   const data = await req.json();
 
-  const userDoc = {
-    _type: "user",
-    _id: user.id,
-    clerkId: user.id,
+  const docId = await ensureUser({
+    id: user.id,
     email: user.primaryEmailAddress?.emailAddress,
     fullName: data.fullName,
-    role: "user",
-  };
+  });
 
   const profileDoc = {
     _type: "profile",
-    user: { _type: "reference", _ref: user.id },
+    user: { _type: "reference", _ref: docId },
     handle: data.handle,
     bio: data.bio,
     jobTitle: data.jobTitle,
@@ -38,7 +28,6 @@ export async function POST(req: NextRequest) {
     location: data.location,
   };
 
-  await client.createIfNotExists(userDoc);
   const created = await client.createIfNotExists(profileDoc);
 
   return NextResponse.json({ profile: created });
@@ -50,15 +39,21 @@ export async function GET() {
     return new NextResponse("Unauthorized", { status: 401 });
   }
 
+  const docId = await ensureUser({
+    id: user.id,
+    email: user.primaryEmailAddress?.emailAddress,
+    fullName: user.fullName,
+  });
+
   let profile = await client.fetch(
     '*[_type=="profile" && user._ref==$id][0]',
-    { id: user.id }
+    { id: docId }
   );
 
   if (!profile) {
     profile = await client.create({
       _type: "profile",
-      user: { _type: "reference", _ref: user.id },
+      user: { _type: "reference", _ref: docId },
     });
   }
 
@@ -72,15 +67,21 @@ export async function PUT(req: NextRequest) {
   }
 
   const data = await req.json();
+  const docId = await ensureUser({
+    id: user.id,
+    email: user.primaryEmailAddress?.emailAddress,
+    fullName: user.fullName,
+  });
+
   const profileId = await client.fetch(
     '*[_type=="profile" && user._ref==$id][0]._id',
-    { id: user.id }
+    { id: docId }
   );
 
   if (!profileId) {
     const created = await client.create({
       _type: "profile",
-      user: { _type: "reference", _ref: user.id },
+      user: { _type: "reference", _ref: docId },
       ...data,
     });
     return NextResponse.json({ profile: created });
