+++ app/sanity/client.ts
@@ -4,5 +4,6 @@ export const client = createClient({
   projectId: "8n5iznjt",
   dataset: "production",
   apiVersion: "2025-06-09",
+  token: process.env.SANITY_API_TOKEN,
   useCdn: false,
 })
